rename ToggleButton to Switch
