+++ packages/react-components/src/components/Switch/Switch.css
@@ -1,0 +1,2 @@
+.bcds-react-aria-Switch {
+}
