+++ packages/react-components/src/components/Switch/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./Switch";
+export type { SwitchProps } from "./Switch";
