--- packages/react-components/src/components/ToggleButton/ToggleButton.css
@@ -1,2 +1,0 @@
-.bcds-react-aria-ToggleButton {
-}
