--- packages/react-components/src/components/ToggleButton/ToggleButton.tsx
@@ -1,17 +1,0 @@
-import {
-  ToggleButton as ReactAriaToggleButton,
-  ToggleButtonProps as ReactAriaToggleButtonProps,
-} from "react-aria-components";
-
-import "./ToggleButton.css";
-
-export interface ToggleButtonProps extends ReactAriaToggleButtonProps {}
-
-export default function ToggleButton({ ...props }: ToggleButtonProps) {
-  return (
-    <ReactAriaToggleButton
-      className={`bcds-react-aria-ToggleButton`}
-      {...props}
-    />
-  );
-}
