--- packages/react-components/src/components/ToggleButton/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./ToggleButton";
-export type { ToggleButtonProps } from "./ToggleButton";
