+++ packages/react-components/src/pages/Switch/Switch.tsx
@@ -1,0 +1,9 @@
+import { Switch } from "@/components";
+
+export default function SwitchPage() {
+  return (
+    <>
+      <Switch />
+    </>
+  );
+}
