+++ packages/react-components/src/pages/Switch/index.ts
@@ -1,0 +1,3 @@
+import SwitchPage from "./Switch";
+
+export default SwitchPage;
