--- packages/react-components/src/pages/ToggleButton/ToggleButton.tsx
@@ -1,9 +1,0 @@
-import { ToggleButton } from "@/components";
-
-export default function ToggleButtonPage() {
-  return (
-    <>
-      <ToggleButton />
-    </>
-  );
-}
