--- packages/react-components/src/pages/ToggleButton/index.ts
@@ -1,3 +1,0 @@
-import ToggleButtonPage from "./ToggleButton";
-
-export default ToggleButtonPage;
