+++ packages/react-components/src/stories/Switch.mdx
@@ -1,0 +1,26 @@
+{/* Switch.mdx */}
+
+import {
+  Canvas,
+  Controls,
+  Meta,
+  Primary,
+  Source,
+  Story,
+  Subtitle,
+} from "@storybook/blocks";
+
+import * as SwitchStories from "./Switch.stories";
+
+<Meta of={SwitchStories} />
+
+# Switch
+
+<Subtitle>TKTK</Subtitle>
+
+<Source
+  code={`import { Switch } from "@bcgov/design-system-react-components"; `}
+  language="typescript"
+/>
+
+## Usage and resources
