--- packages/react-components/src/stories/ToggleButton.mdx
@@ -1,26 +1,0 @@
-{/* ToggleButton.mdx */}
-
-import {
-  Canvas,
-  Controls,
-  Meta,
-  Primary,
-  Source,
-  Story,
-  Subtitle,
-} from "@storybook/blocks";
-
-import * as ToggleButtonStories from "./ToggleButton.stories";
-
-<Meta of={ToggleButtonStories} />
-
-# Toggle Button
-
-<Subtitle>TKTK</Subtitle>
-
-<Source
-  code={`import { ToggleButton } from "@bcgov/design-system-react-components"; `}
-  language="typescript"
-/>
-
-## Usage and resources
